backend/requirements.txt
flask == 3.0.3
flask-SQLAlchemy == 3.1.1 
flask-marshmallow == 1.2.1
marshmallow-sqlalchemy == 1.1.0
flask-cors == 5.0.0
PyMySQL == 1.1.1